Add OAB description

+++ src/OrbitalSurvey/Managers/Core.cs
@@ -81,7 +81,7 @@ public class Core : MonoBehaviour
             SaveManager.Instance.LoadData();
     }
 
-    public void DoScan(string body, MapType mapType, double longitude, double latitude, double altitude, double scanningCone)
+    public void DoScan(string body, MapType mapType, double longitude, double latitude, double altitude, float scanningCone)
     {
         // Sometimes, load data can be done before the textures are initialized
         if (!MapsInitialized)
